+++ sdk/cosmos/azure-cosmos/azure/cosmos/_service_request_retry_policy.py
@@ -18,9 +18,9 @@ class ServiceRequestRetryPolicy(object):
         self.args = args
         self.global_endpoint_manager = global_endpoint_manager
         self.total_retries = len(self.global_endpoint_manager.location_cache.read_regional_endpoints)
+        self.total_in_region_retries = 2
         self.in_region_retry_count = 0
         self.failover_retry_count = 0
-        self.total_retries = 1
         self.connection_policy = connection_policy
         self.request = args[0] if args else None
         self.logger = logging.getLogger("azure.cosmos.ServiceRequestRetryPolicy")
@@ -37,8 +37,6 @@ class ServiceRequestRetryPolicy(object):
         :returns: a boolean stating whether the request should be retried
         :rtype: bool
         """
-        # Fed A
-        #
         if not self.connection_policy.EnableEndpointDiscovery:
             return False
 
@@ -48,44 +46,73 @@ class ServiceRequestRetryPolicy(object):
             if self.request.operation_type == ResourceType.DatabaseAccount:
                 return False
 
-        # We just failed on regional endpoint, let's mark it unavailable and send the cache refresh flag
-        # Cache refresh flag should only be true if we have already tried both current and previous regional endpoints
-        if self.request:
-            # This refresh cache logic is limited to only 2 regional endpoints being stored in location cache.
-            # This needs to be changed once we start receiving multiple regional endpoints for a single region.
             refresh_cache = self.request.last_routed_location_endpoint_within_region is not None
-            if self.location_endpoint:
-                if _OperationType.IsReadOnlyOperation(self.request.operation_type):
-                    self.global_endpoint_manager.mark_endpoint_unavailable_for_read(self.location_endpoint, refresh_cache)
-                    self.logger.warning("Marking %s unavailable for read", self.location_endpoint)
-                else:
-                    self.global_endpoint_manager.mark_endpoint_unavailable_for_write(self.location_endpoint, refresh_cache)
-                    self.logger.warning("Marking %s unavailable for write", self.location_endpoint)
+            self.mark_endpoint_unavailable(refresh_cache)
 
-            self.in_region_retry_count += 1
-            self.request.last_routed_location_endpoint_within_region = self.request.location_endpoint_to_route
-            # The reason for this check is that we retry on
-            # current and previous regional endpoint for every region before moving to next region
-            if self.in_region_retry_count > 1:
+            # Check if it is safe to do another retry
+            if self.in_region_retry_count >= self.total_in_region_retries:
+                # reset the in region retry count
                 self.in_region_retry_count = 0
-                self.failover_retry_count += 1
-                self.request.last_routed_location_endpoint_within_region = None
+                if (self.failover_retry_count + 1) >= self.total_retries:
+                    return False
 
-            if self.failover_retry_count >= self.total_retries:
-                return False
+            self.request.last_routed_location_endpoint_within_region = self.request.location_endpoint_to_route
 
-        if self.request:
-            # clear previous location-based routing directive
-            self.request.clear_route_to_location()
+            if _OperationType.IsReadOnlyOperation(self.request.operation_type):
+                # We just directly got to the next location in case of read requests
+                # We don't retry again on the same region for regional endpoint
+                self.location_endpoint = self.resolve_next_region_service_endpoint()
+            else:
+                self.location_endpoint = self.resolve_current_region_service_endpoint()
+                # This is the case where both current and previous point to the same writable endpoint
+                # In this case we don't want to retry again, rather failover to the next region
+                if self.request.last_routed_location_endpoint_within_region == self.location_endpoint:
+                    # Since both the endpoints (current and previous) are same, we mark them unavailable
+                    # and refresh the cache.
+                    self.force_refresh_cache()
+                    # Although we are not retrying again in this region
+                    # but since this is the same endpoint, our in region retries are done for this region
+                    # and we reset the in region retry count
+                    self.in_region_retry_count = 0
+                    # Check if it is safe to failover to another region
+                    if (self.failover_retry_count + 1) >= self.total_retries:
+                        return False
+                    self.location_endpoint = self.resolve_next_region_service_endpoint()
 
-            # set location-based routing directive based on retry count
-            # ensuring usePreferredLocations is set to True for retry
-            self.request.route_to_location_with_preferred_location_flag(self.failover_retry_count, True)
-
-            # Resolve the endpoint for the request and pin the resolution to the resolved endpoint
-            # This enables marking the endpoint unavailability on endpoint failover/unreachability
-            self.location_endpoint = self.global_endpoint_manager.resolve_service_endpoint(self.request)
-            if self.location_endpoint is None:
-                return False
             self.request.route_to_location(self.location_endpoint)
         return True
+
+    # This function prepares the request to go to the second endpoint in the same region
+    def resolve_current_region_service_endpoint(self):
+        # clear previous location-based routing directive
+        self.request.clear_route_to_location()
+        # resolve the next service endpoint in the same region
+        # since we maintain 2 endpoints per region for write operations
+        self.request.route_to_location_with_preferred_location_flag(self.failover_retry_count, True)
+        return self.global_endpoint_manager.resolve_service_endpoint(self.request)
+
+    # This function prepares the request to go to the next region
+    def resolve_next_region_service_endpoint(self):
+        # This acts as an index for next location in the list of available locations
+        self.failover_retry_count += 1
+        # clear previous location-based routing directive
+        self.request.clear_route_to_location()
+        # clear the last routed endpoint within same region since we are going to a new region now
+        self.request.clear_last_routed_location()
+        # set location-based routing directive based on retry count
+        # ensuring usePreferredLocations is set to True for retry
+        self.request.route_to_location_with_preferred_location_flag(self.failover_retry_count, True)
+        # Resolve the endpoint for the request and pin the resolution to the resolved endpoint
+        # This enables marking the endpoint unavailability on endpoint failover/unreachability
+        return self.global_endpoint_manager.resolve_service_endpoint(self.request)
+
+    def mark_endpoint_unavailable(self, refresh_cache: bool):
+        if _OperationType.IsReadOnlyOperation(self.request.operation_type):
+            self.global_endpoint_manager.mark_endpoint_unavailable_for_read(self.location_endpoint, True)
+            self.logger.warning("Marking %s unavailable for read", self.location_endpoint)
+        else:
+            self.global_endpoint_manager.mark_endpoint_unavailable_for_write(self.location_endpoint, refresh_cache)
+            self.logger.warning("Marking %s unavailable for write", self.location_endpoint)
+
+    def force_refresh_cache(self):
+        self.global_endpoint_manager.force_refresh(None)