+++ sdk/cosmos/azure-cosmos/azure/cosmos/http_constants.py
@@ -256,7 +256,7 @@ class HttpHeaders:
     ThinClientProxyOperationType = "x-ms-thinclient-proxy-operation-type"
     ThinClientProxyResourceType = "x-ms-thinclient-proxy-resource-type"
 
-    # Client Id header for load balancing
+    # ClientId header for load balancing
     ClientId = "x-ms-client-id"
 
 class HttpHeaderPreferenceTokens: